requirements.txt
configparser==3.5.0
Django==1.10.1
django-cors-headers==1.2.1
djangorestframework==3.4.7
httpie==0.9.6
mysqlclient==1.3.9
Pygments==2.1.3
requests==2.11.1
